LPS-147890 Move logic to JavaUpgradeProcessCheck.java
